chore: remove old EmptyState component
