+++ src/components/common/index.ts
@@ -1,6 +1,5 @@
 export { DataTable } from "./DataTable";
 export { LoadingSpinner } from "./LoadingSpinner";
-export { EmptyState } from "./EmptyState";
 export { ConfirmDialog } from "./ConfirmDialog";
 export { FormField } from "./FormField";
 export { MiniSparkline } from "./MiniSparkline";
